Move crossbow to archery
